Change 'doimagemask' so that the SMask image is not processed twice when the device is 'pdfwrite'. Also fix the 'freed SMask' issue of 690345 since the first improvement results in a dereference of this that did previously was only detected during Valgrind.
DETAILS:

The pdf_image3x_make_mcde function now does not set  SMask because
pmie->writer.pres->object may be substituted after the image stream is
accummulated. pdf_end_and_do_image will set SMask with the right value.
Setting this to early ended up with the Pattern Resource having a
pointer to freed memory seen with pdfwrite and the file from 688728.

The changes in pdf_draw.ps reverse the sense of the test for 'pdfwrite'
in doimagemask (for readability) and the important thing -- pop the
SMask image dict when it is to be skipped (this was missing in the
previous change to this area).

These changes are expected to fix the previous problem files seen with
pdfwrite on this smask_work branch, typecheck errors and other PS
errors.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9662 a1074d23-0009-0410-80fe-cf8c14f379e6