Start of removal of Argyll ICC dependencies.
DETAILS:

For now, I have just ifdef'ed and commented out the dependencies as they are conflicting with littlecms as I try to hook it into place.  Argyll will eventually be brought back in as an optional CMS.  Much of the commented out code will be replaced but I did not want to delete it at this time.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@9671 a1074d23-0009-0410-80fe-cf8c14f379e6