Allow inline images in pdfwrite.
DETAILS:

This had been disabled during the softmask branch work.  With the fix that Ray did to avoid having softmask images to be processed outside the 3X code with an extra softmask group, we can re-enable this.  From a test that Ken did we need the ability to do in-line to avoid creating large files in certain cases with pdfwrite.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9675 a1074d23-0009-0410-80fe-cf8c14f379e6