Allocate the inferno device's color cube dynamically.
DETAILS:

The inferno device generates a color cube on device open, ostensibly
to speed up the palette mapping. Previously this was stored in a
global array, adding 16 or 32 KB to the constant memory footprint
of ghostscript.

Instead, this commit allocates the color cube dynamically, storing
a pointer to it in the inferno_device structure. A custom structure
descriptor is added, and propagated through the new prn_device
construction macros, so it can be traced by the garbage collector.

In this way, the memory is only used when the inferno device is active.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9684 a1074d23-0009-0410-80fe-cf8c14f379e6