Fix (pdfwrite) : Enlarge the hardcoded threshold for subsetting True Type fonts.
DETAILS :
Patch and log supplied by Leonardo.

Bug 689236 "pdfwrite disregards -dSubsetFont= with TrueType".

This fixes the valid part of the bug, which is related
to not-subsetting True Type fonts. The supplied example
exceeds the hardcoded constant of 500 glyphs,
so the font was always subset. Here we enlarge
the threshold to a more reasonable value. Note it doesn't
affect regular runs with "standard" distiller parameters.

ref# leo2079.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9696 a1074d23-0009-0410-80fe-cf8c14f379e6