Fix (FAPI): Incremental Type 42 font with GlyphDirectory & missing glyphs fails
Details
The FAPI code assumed that the routine to extract glyph data from sfnts would always
work when a GlyphDirectory is present, and used glyf + loca if this routine failed,
assuming that this meant there was no GlyphDirectory. However, the test file
Bug687845.ps contains a type 42 incrementally downloaded font which uses a
GlyphDirectory and also attempts to access glyphs which aren't defined in the
GlyphDirectory.

This caused the code to try and use the loca and glyf tables, which aren't present,
which caused an error.

Modified the FAPI code to return a glyph length of 0 when we have a GlyphDirectory but
no defined glyph. This causes GS to render nothing, which matches the regular result as
well as the action of Acrobat.

Expected Differences
Bug687845.ps should now work with FreeType

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9714 a1074d23-0009-0410-80fe-cf8c14f379e6