Fix for case when blending color space of soft mask is not specified.
DETAILS:

When the blending color space for the soft mask is not specified, the blending color space is set to UNKNOWN.  This is later resolved when the compositor action is handled.  In this case, the parents color space or the device color space is used for blending.  The code was in place
to do this for a transparency group that did not have a defined color space but not for a soft mask. This caused bug 690457 with gxps which should now be resolved.

EXPECTED DIFFERENCES:

No differences on current PDF test files.  XPS test files that have opacity masks should no longer seg fault.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9715 a1074d23-0009-0410-80fe-cf8c14f379e6