Fix (FAPI) : FAPI fails when FontBBox is missing
Details
(see also rev 9692) The FAPI code attempts to modify the FontBBox in the font
dictionary after rebuilding a font with a non-GS font renderer. If the FontBBox is
missing this creates an error condition. This is generally correct, FontBBox is a
mandatory entry in a font dictionary.

However, the code can be called to define a (PDF) type1C font, a CFF font, and it is
possible (Bug689740.pdf) that the top DICT does not contain a FontBBox. In this case
we should inherit the FontBBox from the FontDescriptor in the PDF file, but we don't
have a mechanism to do that.

Since PDF isn't a programming language the scope for using the FontBBox is pretty
limited, so for now I've chosen to ignore the error.

Expected Differences
None (bug689740.pdf should now work with FreeType though)

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9725 a1074d23-0009-0410-80fe-cf8c14f379e6