Fix : Windows drag&drop has a fixed path limit of 80 characters.
Details
Bug #690461 "ghostscript can't handle files with long name-paths"

The real issue is that the drag-and-drop interface had a fixed buffer of 80
characters, and was unable to process files beyond that limit.

It looks like the original code was from some boiler plate sample. Updated the code
to use the DragQeryFiles API call to find the length of the full path specification
and then malloc a buffer to hold it. Although we would never normally use the system
malloc call, this should not be a problem in the Windows-specific code.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9731 a1074d23-0009-0410-80fe-cf8c14f379e6