Fix regression caused by the changes to make PCL work. Bug 690463.
DETAILS:

The primary culprit was that the 'iblock' calculation didn't work if the
entire 'cbuf' wasn't empty. The token scanner sometimes refilled when there
was still data in the buffer. The 's_block_read_seek' also needed to make
sure that the cursor.r ptr and limit (srptr and srlimit) is set correctly
before loading the block at the new position. Tested with PCL and PS on
the cluster (and of coures on my Windows machine).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9733 a1074d23-0009-0410-80fe-cf8c14f379e6