Use 32 bits for fixed coordinates even on 64 bit cpus.
Details:

The fixed coordinate system used to be dependent on the size of a long.
On x86_64 and other 64 bit systems this was 8 bytes but 4 bytes on other
systems.

Fixes Bug 690434.

Expected Differences:

On 64 bit systems many minor differences; since the nightly regression is run on a 32 bit system no differences.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9734 a1074d23-0009-0410-80fe-cf8c14f379e6