Fix (ps2write) : Crash caused when converting large shading to image.
DETAILS :

Bug 690475 "ps2write" segfaults on PDF file of CUPS test suite".

When converting a shading to an image for PostScript output, and the resulting image is
large (> 512Kb), the code attempts to select an appropriate compression filter. As part
of doign this it probes the colour space in an attempt to determine what kind of space
it is. However this doesn't work when we have created the image ourselves, as tehre is
no associated imager state to probe the colour space resulting in a crash.

Modified the compression selection code to not probe the space if there is no imager
state.

EXPECTED DIFFERENCES :
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9738 a1074d23-0009-0410-80fe-cf8c14f379e6