Fix (FAPI) : FAPI doesn't support Multiple Master fonts
Details
The FAPI code manufactures a stream containing font data for the benefit of FreeType.
However, when manufacturing a type 1 font stream it did not account for the possibility
of a Multiple Master font (or instance of a MM font). This caused FreeType to generate
an error when trying to process glyphs using MM operations.

Extended the existing code to manufacture the WeightVector array, $Blend procedure,
BlendDesignMap array, BlendDesignPositions array and BlendAxisTypes array if present
in the original font.

Expected Differences
None (tpc2.ps now works with FAPI & FreeType)

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9742 a1074d23-0009-0410-80fe-cf8c14f379e6