Provide fallbacks in jbig2dec.c if PACKAGE and VERSION aren't defined.
Normally, the configure script defines PACKAGE and VERSION, based on
its initializers. However, config.h is in fact optional, so we shouldn't
assume these are set. Since we use these to implement the command
'jbig2dec --version', we provide fallbacks in the source file to
simplify compiling outside the autotools build.