Fix the jam build to work without config.h.
Previously, the Jamfile assumed a config.h had be created, usually by
running the configure script, but possibly also created manually. This
is extra work for those wanting a quick build. Therefore we pass some
defines on the compiler command line instead.

These are also likely to be wrong: we assume libpng is available on
all non-windows platforms. It's an incremental improvement, however,
and actual build-time configuration based on detected resources is
difficult in Jam.