Fix an uninitialized variable warning in FAPI_do_char.
Details:

Previously the local variable 'code' was used for incidental return
value checks, but *also* to set cr.is_glyph_index in in the CID font
case. This is because cid_to_TT_charcode() returns 1 on success, but
if it can't find a character code for the given cid, it sets the
character code to zero and returns zero itself to indicate failure.

Zero makes more sense as a glyph id (.notdef) than as a character
code, so in this case it's appropriate to set is_glyph_index based
on this. The other paths which set 'code' all deal with truetype
fonts, so the default success value of zero is also correct.

However, there were in theory code paths where code was never set,
and so is_glyph_index was being set from an unitialized value; thus
the gcc warning. And the double-use of the variable 'code' was hard
to read. This commit adds a separate boolean to track whether we've
found a character code or not and sets cr.is_glyph_index based on
that.

Expected Differences:

None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9760 a1074d23-0009-0410-80fe-cf8c14f379e6