Several memory related fixes.
DETAILS:

Added release method for profile handle.  Added reference counting for profile objects.   Increased initial size of userparms dict in icontext.c .  Added path name to profile.  This can be used as a quick compare in cases where the user params are set multiple times.  We can then avoid releasing and recreating the profiles in the profile manager.

Made change in irclaim.c so that the context_state_load occurs AFTER the chunks are reopened.  We have allocations that can occur during the context state load and we except those to remain valid for later GC.  If memory is closed when they occur, then we do not have the proper top address when we open the memory.

Also changed gsicc_set_profile to use const char and length instead of gs_param_string.  This will make it easier for other interpreters to set the default source color profiles.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@9776 a1074d23-0009-0410-80fe-cf8c14f379e6