Build a shared library with GNU libtool.
Previously we used autoconf and automake, but just build a static
library. This commit checks for libtoolize in autogen.sh and adds
support for libtool to the configure.an and Makefile.am scripts.

We use this to build both static and dynamic versions of the jbig2dec
library, including soname versioning for the later.

Based on a patch by Sebastian Rasmussen.