Manpage updates.
List the file.jbig2 and the embedded global+page stream pair
invocations separately, since they are exclusive.

Give a more complete description and elaborate on some of the options.

Use dot-macros instead of backslash commands for most of the formatting.
This isn't any less confusing because of the need to use double quotes
to enforce whitespace when alternating styles, but I still find it
more reasonable.

Fix a couple of spelling errors.