Handle 4-bit grayscale JPXDecode images.
This is a minimal patch to fix bug 690174.

The file has a JPXDecode stream with 4 bits per component. The image
dictionary has the correct matching /BitsPerComponent and /ColorSpace
keys, but the stream filter was returning the pixel data as the low
nibble in 8 bit pixels.

This is a minimal fix: if the first component of the data is 4 bpc,
divide the stride in half, then in copy_row_gray, pack two pixels
per byte when returning the data.

Expected Differences:

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9801 a1074d23-0009-0410-80fe-cf8c14f379e6