Fix (pdfwrite) : Update 'where_used' in case object is used across pages
Details
Bug #690056 "Error with embedded image in generated PDF: XObject not found"

Pdfwrite tries to detect reuse of objects, and only emit a single copy when this occurs.
However on return from pdf_substitute_resource several places do not set the
'where_used' field. If the object is used on a different page to the one where it is
defined, this causes it not to be listed in the page Resources.

Update two of the places where this occurs to resolve the problem.

Expected Differences
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9816 a1074d23-0009-0410-80fe-cf8c14f379e6