Enhancement (TrueType fonts): Add support for Adobe's use of Identity CMap in PDF files, when substituting a TrueType font, mapped as a CIDFont from disk, for a missing CIDFont.
Details:
bug #688515 "Need method to map CIDFonts to TrueType fonts when Ordering is Identity"
This issue also covers issues #689499, #688813, #689623 and #690483.

When using a native TrueType font from disk as a replacement for a CIDFont, GS needs a
way to convert the character code used in the input file (either PostScript or PDF) into
a TrueType glyph ID (GID).

When the TrueType font is declared (in cidfmap), the CSI entry contains an Ordering, this is concatenated with the CMAP declared in the actual TrueType font. This
information is used to create a CIDDecoding resource, certain  combinations (defined in
.CMapChooser) create CIDDecoding resources 'on the fly'.

A CIDDecoding resource simply maps a character code directly to a GID. Normally this
resource, when automatically created, is manufactured by converting the character code
to an internal code (using the declared CMap from cidfmap) and then converting the
internal code to a GID using the TrueType CMAP table. Most often the internal code
is a Unicode code point.

The investigation in 688515 makes it clear that when we have an Identity CMap in a
PDF file the character code should be treated as the GID. This makes it very easy for
us to construct a CIDDecoding resource, its a simple one-to-one. Unfortunately the
existing machinery relies upon a CMap, and there is already a mapping for
Identity.Symbol (An Identity CMap with a font whose CMAP table is a Symbol table).
We don't want to break this existing functionality, so the new code adds an entry
in .CMapChooser for Identity.Unicode (only to avoid errors) and then in the
function which actually loads the TT font (load_sfnts in gs_cidtt.ps) we check
specifically for an Identity.Unicode Decoding. If we find one then we call a new
routine to create an Identity mapping (CIDMap) from character code to GID.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9834 a1074d23-0009-0410-80fe-cf8c14f379e6