Fix for crash when tiles are outside printable areas seen with bug 690558.
DETAILS:

The 'fit_copy' macro adjusts the x, y, w, and h values, but since this is
a procedural macro and the 'FOR_RUNS' macro (also procedural) declares
local variables, for 'C' the 'FOR_RUNS ... END_RUNS' needs to be in
a block. The only real changes are the additions of 'fit_copy()'.

EXPECTED DIFFERENCES:

Regression testing shows none.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9837 a1074d23-0009-0410-80fe-cf8c14f379e6