Fix (graphics library) : Conversion to base colour space of an /Indexed colour space was not catering for spaces with a string lookup instead of a procedural lookup.
Details
Bug #690587 "setcolorspace + currentgray crashes GS 8.64"

The conversion from a colour in an indexed space to its alternate assumed a procedural
lookup and failed badly if the space had a string lookup.

Expected Differences
Since this is only used when calculating a base colour equivalent (currentgray and
friends), which are rarely used in real world jobs, no differences are expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9841 a1074d23-0009-0410-80fe-cf8c14f379e6