Addition of code for transparency group color space handling when we are filling with transparent tiles and added code for blending with self overlapping pattern tiles
DETAILS:

Addition of code to ensure that the proper color space is used during the push of the transparency group for the pattern path fill.  The base color space for the pattern is used.  When the group is popped it will then be converted and blended with the parent group.  Also, added code so that if the step matrix is such that the tiles overlap we will do proper transparency blending of the tiles.  This code should be optimized at some point.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/pattern_trans@9854 a1074d23-0009-0410-80fe-cf8c14f379e6