Add a couple of utilities for extracting Fonts and ICC Profiles from a PDF. I had these laying around and the need for these comes up from time to time. See the top of each PostScript file for usage.
DETAILS:

It is the responsibility of creators of PDF's to not embed fonts that
are identified as not to be embedded. Ghostscript (as of rev 8603)
respects the TrueType FSType FontInfo.

The copyright and license of fonts and ICC profiles in a PDF will
still be in force. Any knowledgeable user can use Ghostscript
(or many other tools) to extract this information. We provide this
utility mostly to provide the 'source' for GPL'ed fonts as required
by fonts. Fonts are programs and distribution requires distribution
of 'source', so this lets users get the source and (using a font
editor) modify the font. See bug 690533 (if you must).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9858 a1074d23-0009-0410-80fe-cf8c14f379e6