Optimizations when filling with patterns that include transparency.
DETAILS:

There are two optimizations here.  One is that if there is no overlap of the tiles, then we want to go ahead and tile without blending.  So we now detect the case when the tiles will not overlap and if that is the case we do a filling which does not blend the overlapping tiles.  The other optimization, is to use memcpy in the inner loop when we are filling with the pattern buffer for the case where there is no blending.  For this we have a memcpy for the left tile portion, a number for mem copies for the full tiles and then a memcpy for the right tile portion.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/pattern_trans@9865 a1074d23-0009-0410-80fe-cf8c14f379e6