Change the msvc xps build to use a 32 bit color index like the gcc build.
The gcc xps build inherits the color index setting from the pcl
makefile, which currently sets it to 'unsigned long' or 32 bits.
The msvc pcl makefile didn't set it at all, relying on the gs
default, which is also 32 bits. The msvc xps makefile was
overriding that with 'unsigned long long' or 64 bits.

Instead, this commit makes the msvc xps makefile inherit from
the pcl makefile, which now explicitly sets 32 bits. So the
only material change should be that the msvc xps build now
matches the gcc xps build.

XPS allows up to 8 colour components, defined by an ICC profile,
so 64 bits is more correct. We don't currently support that,
pending the completion of the icc_work branch. When support
for multichannel profiles is added the default will need to
be changed for both builds.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9885 a1074d23-0009-0410-80fe-cf8c14f379e6