Fixes a regression with revision 5219, the log for this revision says:
  "Valid memory pointer used create a tile_clip device instead of
  NULL"

Unfortunately the tile clip device is created on the c stack and
should not get a non null memory pointer.  Later in gxcht.c this same
device memory pointer is used to get a gs id, maybe this was the real
motivation for the change.  We now use the device halftone memory
pointer and the device is initialized with a NULL memory pointer as
expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9931 a1074d23-0009-0410-80fe-cf8c14f379e6