Fix for bug 690670.  Softmask color space was incorrectly getting set to DeviceN when the color space was undefined for the group and the output device was a separation device.
DETAILS:

For a transparency group, this method is OK as we blend the individual channels as defined in the PDF spec.  For a mask, this is not OK, since we have to map to luminosity and the mapping from the process + spot colors to luminosity is not defined.  Solution is to use the concrete space of the current space of the imager state for the softmask for this situation.

DIFFERENCES EXPECTED:

None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9935 a1074d23-0009-0410-80fe-cf8c14f379e6