Base the compiler configuration for lcms on $(CC) instead of $(CC_).
Details:

When compiled under MSVC (_MSC_VER defined), lcms configures itself
specially to build under that compiler, including using msvc-syntax
inline assembler and including windows.h for thread safety and
convenient type references. However, both these features are
incompatible with the /Za switch (compile as ANSI C/C++) which
msvccmd.mak applies to normal command lines.

Elsewhere in the build system, the $(CC_WX) config is used for
files that need windows.h, like client applications and ijs_exec_win.c.
This config seems to be used to propagate various flags for the default
calling convention to the borland compiler, but all of the
windows-oriented makefiles reproduce the machinery. It has sense become
a shortcut for a non-ansi version of the compiler for windows-specific
source files.

However, $(CC_WX) isn't defined in any of the non-windows makefiles.
Reverting to just $(CC) could cause problems from missing compiler
flags, especially for the invocation for gsicc_littlecms.c which is
otherwise part of the base library and not an external component,
however I judged this a better solution than propagating a new CC_EXT
config through all the builds. Tested on linux-gcc and win32-msvc.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@9969 a1074d23-0009-0410-80fe-cf8c14f379e6