Start of ICC based color conversion for 8 bit images.  This will apply the transforms on the data as each row is presented to image_render_color.
DETAILS:

The data is already skipped in the y dimension if we have a reduction.  Currently all the data in a full resolution row is transformed.  If there is a significant reduction in X, it will make sense to go ahead and do that reduction first and then transform the color.  This needs to be added.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@9983 a1074d23-0009-0410-80fe-cf8c14f379e6