Remove code that tries to pause at the end of the .PFB binary data in an attempt to keep from reading beyond the end of the encrypted data, sometimes misidentifying the end of the actual data, and stopping early. Bug 465936.
DETAILS:
Since the eexec operator specifies that one never reads more than 256 bytes
ahead, and since the pad bytes are given in the text portion of a .PFB file,
in hexadecimal, there is no need to go to extra lengths to avoid reading
(too far) past the end of the eexec data. Thanks to Paul Vojta and Ray
Johnston for the path.

DIFFERENCES:
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9992 a1074d23-0009-0410-80fe-cf8c14f379e6