Addition of code to allow setting of an icc search directory where the icc profiles specified for defaultgray, defaultrgb etc may be found when those are specified by the user.
DETAILS:

The directory path is stored in the iccmanager object.  This object now has 8 pointers which are GC, so an 8 pointer macro was defined in gsstruct.h .  The directory is specified by a string user parameter given by ICCProfilesDir.  The directory is set by calling gsicc_set_icc_directory from a function in zusparam.c.  A similar type of setting should be done for the other languages.  A search for the profiles was set up to search only in the search directory if one was given.  If one is not specified, then the default files are search for at the gs root level (or the absolute or relative path that was specified by the appropriate command line option for that color space e.g. -sDefaultGrayProfile = "c:\myprofiles\Gray_Default.icc").  If those are not found, an attempt is made to find the profiles in %rom% . Those will only be found if the default profiles are used (i.e. the profile was not specified in the command line).

The motivation for the search directory is to make it easy for the user to specify only the profile file names for all the various profiles, which are usually found in a single directory on most systems or for most users.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@10010 a1074d23-0009-0410-80fe-cf8c14f379e6