Change in PDF interpreter code so that the PDF color spaces Lab, CalGray and CalRGB will be passed as dictionaries with their parameters to the c code that handles color spaces.
DETAILS:

These color spaces will be represented by ICC profiles in the new approach.  Currently in the trunk they are implemented as CIEABC ( CalRBB and Lab) or CIEA (CalGray) where the nonlinearities are put into postscript procedures that are later sampled during the construction of the jointCIE cache.  With the upcoming changes, ICC profiles will be created instead using the parameters (or in the case of LAB using a ready made LAB profile).  These profiles will be linked in with the destination profiles when it is time to transform a color.  The C-code for all of this is coming in different commits.  This commit is only the changes in the PDF interpreter so that we no longer cast these color spaces into CIEDicts.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@10021 a1074d23-0009-0410-80fe-cf8c14f379e6