Changes to ensure that separation devices have their CMYK values color managed while the remaining spot colors pass through unmolested.
DETAILS:

Many users wish to be able to color manage the CMYK process colorants but have the process colorants be unmanaged when going out to separation devices.  That is what occurs with these changes.  When a separation device is used as an output device and no ICC output profile is specified, then the default CMYK profile is used in defining the output color space (this is also the default requirement for XPS).  It is possible to define an CMYK ICC profile of your choice for the device using the -sOutputICCProfile=my_device_profile.icc command line option.  Still to come is being able to set a DeviceN profile to achieve n-color ICC color management with up to 15 colors using an ICC profile.  I don't believe this is common with users, but it should be straight forward to implement.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@10047 a1074d23-0009-0410-80fe-cf8c14f379e6