Setting of default ICC profiles for Gray, RGB and CMYK into DeviceGray, DeviceRGB, and DeviceCMYK color spaces.
DETAILS:

When a Gray, RGB, or CMYK color space is set as the current color space of the graphic state, the appropriate ICC profile is assigned to that color space from the ICC Manager and the type procs for that color space are altered to ICC.  We then apply proper color management to the device input colors.  Todo still is making it clear how to avoid doing color management ( and testing that it works ), which is a common requirement.  That should occur automatically if the source and destination profiles are the same, in which case the link is the identity mapping and no remap will occur.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@10050 a1074d23-0009-0410-80fe-cf8c14f379e6