Fix for bug 690534.
DETAILS:

We now handle the softmask BC entry properly.  This background color is to be composited with the drawings into the softmask.  We now properly initialize the softmask buffer to have this compositing occur.  If we are outside the softmask bounding box then the luminance of the background color is used as the softmask value.  This fixes several rendering issues.

DIFFERENCES EXPECTED:

Bug690027.pdf  PROGRESSION
Bug688807.pdf  PROGRESSION
Bug690170a.pdf PROGRESSION
Bug690170.pdf  PROGRESSION

Bug690022.pdf  I would like a second opion on this one.  It looks like there may be a minor regression.  I am doing the commit anyway due to the large number of progressions with this fix.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10063 a1074d23-0009-0410-80fe-cf8c14f379e6