Remove an unncessary keyword substitution flags from the Adobe CMaps.
These files don't have an RCS/SVN version header, so the substitution
could only cause problems. In the case of the two Identity-UTF16 CMap
files which are owned by Artifex and do have version headers, I've
reduced the substitutions to just the 'Id' keyword to match the rest of
the source.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10073 a1074d23-0009-0410-80fe-cf8c14f379e6