Fix for transparency problem, caused by the alpha parameters in the ExtGState in the PDF interpreter (dictstack) not being maintained properly. Bug 690770 for customer #1110. Thanks to Michael Vrhel for analyzing the PDF to find the area that the PDF interpreter was mishandling.
DETAILS:

The ExtGState is kept as values in the dictionaries on the dictstack and are
sent to the graphics library as needed using '.set***alpha' operators. The
CA and ca values were properly set to 1 in the graphics library when the Do
operator was used to execute Form or Image objects, but the corresponding
FillConstantAlpha and StrokeConstantAlpha values were not being set in the
ExtGState so any operations in a Form that set the values in the graphics
library from the ***ConstantAlpha values in the ExtGState could set them
to incorrect values.

EXPECTED DIFFERENCES:

CATX3755.pdf
CATX3146.pdf
CATX5720.pdf
CATX2144.pdf

for pkm, pbm and ppm giving 12 expected differences.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10088 a1074d23-0009-0410-80fe-cf8c14f379e6