Fix to make sure that we use the proper bounding box when computing the band range for the soft mask transparency group in the clist.
DETAILS:

Previously, the range on the transparency group was always based upon the range of the previous transparency group.  This had problems that led a seg faults or empty transparency buffers when we had softmask and groups embedded within a softmask due to the end groups occurring when there was no begin group in certain bands.  See bug 690777 for an example.  With this change, the band range is properly based upon the softmask transparency group range and that range is pushed into the cropping buffer so that the end group is matched across the bands.  Note that if the softmask transparency group has a BC entry that is significant (non black) it can have an effect outside the softmasks bounding box.  In this case the parent groups bounding box is used for the softmask band size.   This change fixes the seg fault in 690777, but it exposes a problem in the file associated with bug 690115.  This file has extended graphic state soft mask changes of different sizes that occur wrapped in [q Q] commands. During clist rendering, this causes the presence of a softmask in bands that it should not be.  The goal of this branch is to address this problem.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/gs_extendgraphic@10115 a1074d23-0009-0410-80fe-cf8c14f379e6