Fix (pdfwrite) : Improve symbolic font detection
Details
Bug #690767 "Problems with PDFA"

Revision 9944 altered the way that we detect whether a font is symbolic or not. This
involved several changes; fixing existing broken code, adding /.notdef detection and
attempting to check encoded positions of glyphs before deciding a font was non-symbolic.

The encoding position test was incorrect for TrueType fonts, however, leading to some
fonts being needlessly declared as symbolic. This patch moves the encoding test to the
font descriptor assembly, which is the correct place to test it.

While this resolves the customer issue, it is not a complete solution. A truly symbolic
TrueType font would still be written as a symbolic font and when creating PDF/A files
there are strict limitations on the form of TrueType fonts. In particular they must only
contain a single CMAP table. Because we do not completely generate TrueType fonts for
embedding, but copy some tables, we cannot meet this restriction.

This will continue to be a potential problem until the TrueType font embedding code is
rewritten properly.

Expected Differences
'Boxes and Text.pdf' and 'fonts.pdf' now have missing fonts labelled as non-symbolic,
which affects the substitution used and alters the output.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10124 a1074d23-0009-0410-80fe-cf8c14f379e6