Fix (pdfwrite) : incomplete submission #10124
Details
Somehow in the confusion of IP address changes and local cluster regression tests I
managed to only commit a partial fix for issue #690767. The missing line of code is
in this patch.

Expected Differences
None (But there appear to be some in the local cluster tests which I will review)

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10125 a1074d23-0009-0410-80fe-cf8c14f379e6