Fix (pdfwrite) : Undo escapement of PostScript strings when writing XMP metadata for PDF/A
Details
Bug #690471 "Parantheses escaped in xmp metadata"

Technically this is covered in gs/doc/ps2pdf.htm under -sDSCEncoding which states that
the default behaviour is simply to copy the data directly to the file with no
translation.

However it seems unreasonable not to at least undo the PostScript escapement which we
must have in order to set the PostScript string and which will cause PDF/A
verification failures if we don't.

Modified the default case (which does not convert to Unicode) to undo the escapement
only.

Expected Differences
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10142 a1074d23-0009-0410-80fe-cf8c14f379e6