Front end changes to coordinate the transparency state flags with the graphic state situation and to have the appropriate compositor actions sent to the pdf14 device in a minimal form (that is send them only when absolutely needed as opposed to polluting the compositor with every q or Q event).
DETAILS:

Two flags were added to the imager state.  One is referred to as xstate_pending.  The other is referred to as xstate_change.  They are wrapped in a structure named trans_flags.  The purpose of these two flags is to detect an extended graphic state push (q) followed by the setting of a graphic state with a softmask followed by extended graphic state restore (Q).

Added zbeginextendedgstate and zendextendedgstate commands in the c code.  These operations respectively execute a gs_push_transparency_state and a gs_pop_transparency_state during the clist writer process.

If we execute a gs_push_transparency_state, which occurs when q occurs in the pdf interpreter, we set xstate_pending to true.  If a Q occurs, xstate_pending will be set back to its previous state, due to the grestore.

If xstate_pending is true and we have a graphic state set that includes a transparency mask, we will eventually obtain an end transparency softmask command.  At that point, the xstate_change flag will be set to true to indicate that we had a softmask change in the graphic state.  If we execute a q, at this point (when the xstate_change is set to true), we will need to send a compositor action PUSH_TRANS_STATE to the pdf14 compositor. When this command is encountered, it will do a push of the current mask_buffer into the mask_buffer stack.

When we excute a /Q operation, we will get to gs_pop_transparency_state.  If the xstate_change flag is set, we know that the soft mask was changed and is bracketed by a q Q operation.  We then send a new compositor action POP_TRANS_STATE to the pdf14 compositor.  This will end up doing a pop of the current mask_buffer so that we will have the correct mask buffer, when we do any further drawing.

Next we need to make the mask_buffer a stack and implement the compositor actions.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/gs_extendgraphic@10153 a1074d23-0009-0410-80fe-cf8c14f379e6