Don't disable dprintf() and friends in non-debug builds.
This reverts a patch from Ray (r2812 in the old ghostpcl repository)
which was a quick hack to silence the extensive tracing output ghostxps
produced at the time. Now most of those calls have been moved to
xps_debug_foo() functions which aren't called in normal builds and
a few warnings which shouldn't particularly be silenced.

The patch worked by just defining the dpf macro to the empty string, but
that creates 'operand of comma operator has no effect' warnings on gcc
4.4 when the remaining arguments tuple is interpreted as a discarded
expression.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10170 a1074d23-0009-0410-80fe-cf8c14f379e6