Print a warning when invoking the 'so' unix makefile targets from pmake.
Details:

The make included in freebsd 7.0 appears to have trouble matching or
collapsing target paths of the form ./bin/../sobin/gsc as GSSOC_XENAME
expands to in the recursive shared build make invocation. GNU make works
fine on the same platform.

Note that we use the @ prefix to hide the command output, which may also
be a gmake-ism, but it works on freebsd 7.0.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10180 a1074d23-0009-0410-80fe-cf8c14f379e6