Have the scons build automatically determine the version string.
SConstruct tries to call 'git describe --tags' which produces a version
string based on the most recent tag, with the number of commits and
current commit hash appended if HEAD isn't a tagged revision.

If that command fails (i.e. building from a release tarball or another
version control system) it attempts to read the CHANGES file and uses
the first version number there.