Addition of code to set DeviceN ICC profiles in the ICC manager.  For ghostscript (PS and PDF) this is achieved from the command line where you can provide a string of ICC profiles file names.  For example -sDeviceNProfiles="6channel.icc; 5channel.icc; duotone.icc"   .  You can use , or ; for a delimiter and leading and trailing spaces are removed.  The profiles are added to a linked list member of the icc_manager called device_n.  Other PDLs will need to call into gsicc_set_profile to set a DeviceN profile.
DETAILS:

PDF does not internally (within the document package) provide ICC support for DeviceN input color spaces (XPS does).  To work around this, we provide the capability to externally define an ICC profile for DeviceN colorspaces within a PDF.  Association of the profile with the DeviceN color spaces is made by storing the spot names in a private tag entry in an NCLR type ICC profile.  Then when we need to map the DeviceN color to device colors, a check is made to see if we have a DeviceN profile for this space in the ICC manager.  If one is found we use it.  If not, then we will use the alternate transform if the device does not understand the colorants directly.

Note this is not to say that DeviceN color management can only be achieved with ICC profiles.  If a developer has a proprietary mixing model for inks, they will be able to hook in their method in the location in the code where the DeviceN colors are processed.

With this commit we have the capability to set the DeviceN profiles in the ICC manager.  The next step is to implement the portion of the code that actually use the profiles.  Complete implementation will require the writing of a small application to add the private tag information into the ICC profiles so that we can perform the association.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@10248 a1074d23-0009-0410-80fe-cf8c14f379e6