flush stream data in close_device().
For a whole set of postscript files (the ps3cet set), the close_device()
finalizer code is called before the PXL output is flushed. This causes
the PJL job footer to be outputted in the middle of the last portion of
PXL output, thus corrupting the PXL data. Adding a flush fixes this problem.
Fixes one of the many problems in bug 690858.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10264 a1074d23-0009-0410-80fe-cf8c14f379e6