This is the initial commit of the 'tiffsep1' device code that emits separation files as 1-bit per pixel TIFF files, a separate file for each separation. This is a CMYK plus Spot color device, and the number of separations (spot colors) supported depends on the gdevdevn.h setting of USE_COMPRESSED_ENCODING. By default this is true as long as the sizeof(gx_color_index) is >= 8, so up to GX_DEVICE_COLOR_MAX_COMPONENTS layers can be generated (current default 64). Without compressed color encoding, then only 8 total (C, M, Y, K and 4 spot colors) can be supported.
Files are named with the separation layer in ( ), e.g. x(Black).tif if
the OutputFile is 'x.tif'. Since no composite CMYK is generated, the
base file is deleted ('x.tif' in the example).

DETAILS:

This device uses a hook of the 'fill_path' device call to collect the
threshold arrays from the imager state. In case a file does not perform
a 'fill' operations, the gsdevice.c code does a dummy fill of an empty
path to allow the 'sep1_fill_path' code to convert the 'order' into
threshold arrays.s

The 'open' and 'put_params' device functions are modified to make sure
that if the clist device is being used, the 'sep1_fill_path' proc will
be re-hooked (the clist initialization overwrites some procs including
the 'fill_path' proc). The sep1_fill_path then forwards to whatever
fill_path proc was set by the gdev_prn_open or put_params.

A later commit will move the 'thresholds_from_order' function to another
common file (probably gsht1.c) and change it to generate a threshold
array of type 'gs_threshold_halftone' with NULL transfer and closure
functions. This function is in gdevtsep.c now since it is the only
device that needs it (for now) and development was easier with all
changes more or less localized (at least non-trivial ones).

The make files are modified so that the 'tiffsep1' device is included
by default (where the tiffsep device is included).

The actual halftoning is done using 32-bit writes of the dithered line
since testing showed this to be 27% faster (although more complex).

Note that the compressed color encoding (the current default) is about
half the throughput compared to direct encoding. Users that are sensitive
to performance (who isn't) and that only need 8 or less total colors
(CMYK plus 4 spot colors) will want to turn compressed color encoding
off.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10284 a1074d23-0009-0410-80fe-cf8c14f379e6