Adobe Acrobat considers the use of clip operation on new path a no-op. Check this and don't clip when there's no path. Bug 690891.
DETAILS:
Besides the change in clipping, this patch reverts the code that preserves
current point position after fill operations and interferes with new path
detection. Reverted code predates source control and its purpose is not clear.
Perhaps, it was an attempt to avoid nocurrentpoint errors, which are now
trapped and handled.

DIFFERENCES:
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10293 a1074d23-0009-0410-80fe-cf8c14f379e6