Fix (pdfwrite) : type 0 fonts with type 3 descendants could confuse pdfwrite by using stringwidth
Details
Bug #690875 "Ghostscript locks up writing PDF file"
There is code in gdev_pdf_text_begin which skips caching of type 3 glyphs when using
stringwidth. This prevents glyphs becoming bitmapped and ensures that pdfwrite and
the glyph cache are coherent.

The code did not check for composite fonts using type 3 descendants. In this case if a
job used some glyphs for stringwidth, then later used at least one of those glyphs and
one glyph not previously encountered in a single string pdfwrite would enter an
endless loop.

pdfwrite would notice that a glyph was not marked as 'used' in its internal cache, and
exit pdf_text_process to run the outline for the first glyph. GS would note that the
first glyph was already cached (from the stringwidth) and move on to the next glyph.
GS would render that glyph, pdfwrite would capture it and mark that character code as
used.

On returning to pdf_text_process the first glyph should be marked as used now, but it
was not, so pdfwrite would once again exit and the loop would repeat.

Fixed by checking string using composite fonts to see if any glyph used a type 3
descendant, and disabling caching if it does.

Expected Differences
None, this is an incredibly unusual condition.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10297 a1074d23-0009-0410-80fe-cf8c14f379e6