purge cached palette whenever color space changes.
The current pxlmono/pxlcolor code caches colorspace and palette calls
and not re-transmitting them if it hasn't changed since the last time
they were set. Whenever the color space changes, the current palette
should be purged, and a new palette should be (re-)transmitted to
the printer the next time an indexed colored image is sent.
This fixes bug 690859.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10302 a1074d23-0009-0410-80fe-cf8c14f379e6