caching SetPen/SetBrush Null calls
For some test files (the PS3 CET set), the pxlcolor/pxlmono code
emits a lot of setpen/setbrush nulls. There is no need to set it again
if the printer already got a previous setpen/brush null. So this patch
caches the null/non-null state of both pen and brush in the
driver and not re-issue set nulls if the last set was a null.

This only gives about 5% file size saving for the test file I wrote it
for. Part of the fix to bug 690857.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10311 a1074d23-0009-0410-80fe-cf8c14f379e6