partly reverting r8576, and other file-size improvements.
After r10313, it becomes possible to revert r8576 (which was
a work-around to bug 688992, but increases file-sizes in some cases).
It is also found that r8576 is in fact more efficient for
small rasters; so it is only partly reverted. In two other places where
write_image_data() is called, switching to gx_default_copy_* for single-line
images/masks is likely to work well also.

This fixes bug 690857, where 12-07C.PS uses a lot of 1-bit images,
and shrinks the file size by a factor of 4.

There are a total of 4 places where write_image_data() is called
(copy_mono/copy_color/fill_masks/image_write_rows) - the 4th does
not use the data_bit field, nor is it likely to be called
with small rasters; so it is left unmodified.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10315 a1074d23-0009-0410-80fe-cf8c14f379e6