Fix (TT fonts) : regression with TrueType fonts
Details:
Bug #690642 "conversion to PDF aborts after 800 pages"
Bug #690921 "Regression: /invalidaccess in --run-- starting with r9897"

The patch for #690642 was an interim fix which was known to have flaws and was supposed
to be revisited to fix them. Unfortunately a file appeared which exhibited an attempt to
put a key in a read-only dictionary before this was done.

This patch resolves both issues by removing the need for a temporary key (hence we no
longer write into a read-only dictionary) and also removes some ugliness (and slight
performance penalty) by not creating a dict from the stack, patching the dict, and
loading its contents back to the stack.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10329 a1074d23-0009-0410-80fe-cf8c14f379e6