Fix (FAPI) : CIDFonts not working with fonts containing a CDevProc
Details
When calling zchar_set_cache we always used the glyph name as one of the arguments,
this argument was later pushed onto the stack for use by a CDevProc if required. The
problem is that with CIDFonts the name is always a null object and we should actually
be supplying the CID to the procedure.

Modified the call to zchar_set_cache when we have a CIDFont to supply the CID instead
of the name

Expected Differences
PDF files containing CIDFonts, and any PostScript files with CIDFonts using a CDevProc
should now work with the FAPI interface.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10387 a1074d23-0009-0410-80fe-cf8c14f379e6