Fix (FAPI) : Incorrect GID supplied to font interpreter for CIDFonts with TT outlines
Details
CIDFonts use a CID, for CIDFonts with TrueType outlines the font contains a CIDToGIDMap
entry which is used to convert the CID into a TrueType Glyph ID (GID). We were not using
the data, which resulted in us passing the CID directly to the font interpreter, which
caused the wrong glyph to be extracted, or failures to occur because the GID was larger
than the number of glyphs in the font.

The type 42 font code in GS converts the CIDToGIDMap into a CIDMap entry, we now use
this to convert the CID into a correct GID for the font interpreter to use.

Expected Differences
Some files will no longer produce errors with FAPI, other files now produce correct
glyphs.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10388 a1074d23-0009-0410-80fe-cf8c14f379e6