Fix (FAPI) : intermittent crashes with TrueType fonts
Details
The sfnts reader code in the FAPI module maintains a pointer to a string from an array
of strings (the /sfnts array). It then used the offset member to point along the string
to the current byte within it. When reading multiple bytes, it checks there is
sufficient data by checking ptr+offset against the size of the string. However the code
did not increment the offset member after reading multiple bytes, instead it incremented
the actual pointer. This completely bypassed the size checking, which could cause a
crash depending on the memory layout. This was only likely to happen with fonts with
very large TrueType tables (ie those not conforming to Type 42 recommendations
regarding string sizes, see tech note 5012)

Altered the code to always increment the offset rather than the pointer so that size
checking works properly.

Expected Differences
A number of intermittent crashes should be resolved when using the FAPI and FreeType

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10389 a1074d23-0009-0410-80fe-cf8c14f379e6