Fix (FAPI) : segmentation violation with Multiple Master fonts
Details
When extracting the $Blend procedure from a Multiple Master font, if the procedure
contained a name object, there was a typo which used an uninitialised variable
instead of the extracted name element to convert the name to a string.

Fixed by using the correct name object instead

Expected Differences
FAPI should now work with Multiple Master fonts.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10398 a1074d23-0009-0410-80fe-cf8c14f379e6