Fix for segfault when image 'clues' are accessed during a GC after a restore has erased (set to zero) the contents. Bug 690915.
DETAILS:

As described in gx_image_free_enum comment, the penum may not be freed if
it is at a higher (outer) save level, so the contents (including the
"clues" are zeroed, but the ENUM in image_enum_enum_ptrs (gxipixel.c)
did not check for this. This was indeterministic since it depended on
when the GC scan ran.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10412 a1074d23-0009-0410-80fe-cf8c14f379e6