Factor out the miter checking code into a single function rather than having it repeated 3 times.
Add a test to the extracted code to spot the case where the cdelta's are too
small to be reliably used to calculate the correct miter point. In such cases
revert to using the raw vectors suitably scaled.

Testing reports 69 changes in the regression files, none of which can be
reproduced locally. This fixes bug 690098.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10425 a1074d23-0009-0410-80fe-cf8c14f379e6