Update the clustermonitor cgi script to return regression results.
This makes use of the CGI module to parse an optional 'report' query
parameter, which if present, returns a specific regression result
instead of the default json dashboard status. Current it only supports
svn jobs, not custom runs submitted through the clusterpush script.

Also cleans up the the paths to use a $clusteroot variable for the
prefix.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10426 a1074d23-0009-0410-80fe-cf8c14f379e6