Fix (FAPI) : FAPI code not expecting unsorted LOCA/GLYF tables
Details
The FAPI code needs to copy the GLYF data, to do this it needs to know how long the
data is. The only way to do this is to consult the LOCA table which gives the offset
to each glyph in the GLYF table. Subtracting the start of the glyph from the start of
the next glyph in the GLYF table gives the length.

The existing code assumed that the GLYF (and therefore LOCA) tables was ordered, that is
that glyph 1 followed glyph 0 and so on. There is no such restriction, glyphs can
appear in any order, so the simple subtraction caused incorrect glyph lengths, which
could lead to truncation of the data and caused errors.

We now pick up the offset of the glyph we want from the LOCA table, and then search the
entire table to find the glyph whose offset is nearest (but after) the offset of the
glyph we want.

Expected Differences
Some more files using TrueType or Type 42 fonts will work with FAPI and FreeType.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10433 a1074d23-0009-0410-80fe-cf8c14f379e6