Fix (FAPI) : TT tables > 64Kb not read properly.
Details
If, while reading a TrueType table for serialisation, we ran off the end of an sfnts
string, we switch to the next string, and carry on reading the data. However, we did not
update the destination pointer leading to corruption of the already processed data and
general corruption of the table.

We now update the destination pointer before continuing to write data to it.

Expected Differences
Bug688154.ps and other files with large TT tables should now work correctly with the
FT FAPI code.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10449 a1074d23-0009-0410-80fe-cf8c14f379e6