Fix (pdfwrite) : Unchecked return value causing indeterminisms
Details
Bug #690762 ""
Reported and investigated by Marcos using Valgrind. The return value from a glyph_info
method was not checked before accessing the members of the 'gs_glyph_info' structure. If
an error occurred the fields of the structure were potentially not initialised, leading
to indeterminate results.

Modified to check the return value properly.

Expected Differences
34_all.ps should no longer produce indeterminate output, and some error conditions
with CIDFonts should now be properly signalled.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10474 a1074d23-0009-0410-80fe-cf8c14f379e6