Revert c10520 which causes problems with the MSVC build.
$(CC_) enforces 'ansi' C under msvc32.mak, disabling any Microsoft
extensions. This often causes problems with 3rd party libraries
which try to take advantage of those extensions.

This means that the -fPIC flag is again not propagated in the GCC
build; this is just to unbreak the build until I can test a different
solution which works on both platforms.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10540 a1074d23-0009-0410-80fe-cf8c14f379e6