Fixes bug 691024, the font list is now printed deterministically. Sundry refactoring and cleanup related to printing the font list and loading fonts.  Adds a new scheme for aliasing fonts, as they are loaded, allowing the removal of the font WingSub.ttf.
The pcl font list was printed out by enumerating the unordered font
dictionary, the font dictionary is created at startup by enumerating
all the fonts and placing a header in each entry of the font
dictionary.  The order the entries are added and later enumerated
depended on how the OS enumerated files in the file system.  Of course
we should not depend on any particular ordering with a dictionary, so
the font list now prints the fonts in pjl font number order.

Also WingSub.ttf, the substitute Wingding font can be removed as we've
added a new scheme to do aliasing of fonts that does not require
duplication of fonts.  A resident font may now be aliased by simply
creating a duplicate name in the resident font table, the reason for
this change is to save on rom space now that compiling the fonts into
the executable has become the default configuration.

This commit lacks cohesion and probably should have been done in two
functional commits:(1) deterministic font list and (2) new font
aliasing scheme, unfortunately the interdependency between the changes
made that difficult.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10564 a1074d23-0009-0410-80fe-cf8c14f379e6