Implement symbolic glyph mapping as described in PDF 1.7. Use it for TrueType fonts with symbolic flag set or having no encoding entry. The flag -dIgnoreTTEncoding is removed because symbolic fonts always do this now and non-symbolic fonts should never ignore encoding. Bug 691035, customer 220.
The new spec includes cmap 3,0 in the supported list. We discovered this
before but didn't use cmap 3,0 as a first choice. PDF 1.7 is not clear about
the choice between symbolic and non-symbolic branches. Current algorithm
is based on a few experiments with AR 8.

The patch causes rendering differences in comparefiles/keyboard.pdf
A few undefined characters change from boxes to empty space.
Current rendering of undefined characters in Ghostscript doesn't match
Adobe Acrobat and should be revised.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10575 a1074d23-0009-0410-80fe-cf8c14f379e6