Fix bug #691043 "Vulnerability report : Ghostscript gs_type2_interpret null ptr dereference (Segmentation Fault)". The problem appears to be caused by a corrupt compressed data stream which results in garbage font data being sent to the type 2 and under some conditions the type 1, font interpreters.
As noted these font interpreters are normally coded to be intolerant of faults, and
do not normally perform much error checking. As a result badly formed (or as in this
case, garbage) fonts can cause serious problems.

Although the null dereference is the cause of the error, this is actually caused by
interpreting the data as a font, and decrementing the font instruction pointer below
the bottom of the instruction stack.

The patch here adopts the suggestion from the original bug reporter of checking the
pointer before the dereference, but in addition adds some checking when the
instruction pointer is decremented to see that it does not descend past the bottom of
the stack. This is because tests showed that it was possible to reach this point with
a non-NULL pointer, which was still invalid and caused a segmentation violation.

In addition the free_glyph routine has been modified to check the pgd (pointer to
glyph data) parameter is non-NULL, and that its 'procs' member is non-NULL, before
attempting to call the free procedure. This fault was exposed by selecting the pdfwrite
device.

Although improved this is not a comprehensive fix, there are other locations where the
font instruction pointer may be incremented past the top, or decremented past the
bottom, of the instruction stack. However the FreeType font interpreter correctly
detects the problems without causing a crash, so I don't think we should expend too
much effort on rewriting our current font code to cope.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10590 a1074d23-0009-0410-80fe-cf8c14f379e6