Update the example file. The old file used a single 'charpath pathbbox' to determine the height of the text. It then used the 'scale' operator to draw larger text.
If the path was fixed to pixel boundaries, and the point size was low (default 15
points) this could lead to a small error which was magnified by the cumulative scale
operations.

This led to significant differences when using the FreeType font code, as it seems to
return a hinted path which is bounded by pixel co-ordinates, unlike the regular GS
code.

The altered code scales the font and recalculates the height of the text each time
giving more accurate results. In addition a minimum size has been introduced (50 points)
for the calculation of the text height. Together these result in much closer output.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10591 a1074d23-0009-0410-80fe-cf8c14f379e6