Bug #691044 ""
Another problem caused by a corrupt compressed data stream, resulting in garbage data
being sent to the TrueType interpreter.

Caused when the TT interpreter is given an argument of 0 to the MINDEX operator. Rather
than flag this as an error I've chosen to simply return. This is because it seems that
a value of 0 is actually legal (stupid, but legal), and simply leaves the stack
unchanged.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10602 a1074d23-0009-0410-80fe-cf8c14f379e6