Fix (pdfwrite) Bug #691058 "Incorrect embedding of CIDToGIDMap for CIDFontType 2".
Previously we would embed a CIDToGIDMap for a CIDFontType 2 font (TrueType outlines) if
the font had a non-identity CIDToGIDMap, or if we were producing PDF/A output.

However the spec (PDF 1.7, Section 5.6, "Glyph Selection in CIDFonts') says that we
should *always* embed a CIDToGIDMap if the font is embedded, and never if the font is
not.

This patch modified the CIDToGIDMap production so that we ignore the production of PDF/A
files, and instead check whether the font is being embedded. If it is being embedded
then we create a CIDToGIDMap, if it is not, then we don't.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10608 a1074d23-0009-0410-80fe-cf8c14f379e6