Fix (FAPI - FreeType). The current FAPI implementation for FreeType ignores any vertical advance defined in a glyph. This causes type 1 glyphs defined with an sbw (not hsbw) which have a vertical advance as well as, or instead of, a horizontal advance to not advance the current point correctly.
There is a separate issue with FreeType which does not return the correct vertical
advance which will be addressed in a separate patch.

This patch adds the vertical advance to the metrics structure and uses it as a
parameter to zchar_set_cache to set the horizontal advance in GS for the glyph.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10615 a1074d23-0009-0410-80fe-cf8c14f379e6