Fix (pswrite)
The pswrite device only partially implemented the 'file per page' output model. It
closed the output file and the bbox device, but did not open a new output file or
reopen the bbox device, leading to potential crashes later when the pswrite device
tried to access the bbox device.

This patch reopens the bbox device, and opens a new output file, at the end of every
page when we are outputting one file per page. The problem is that this will emit an
empty EPS after the last page is completed. Fixing this would mean delaying opening the
output file until we had received a marking operation. That's more work than we want
to do on this device, so this is being left as an issue.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10630 a1074d23-0009-0410-80fe-cf8c14f379e6