Added a new option to clusterpush.pl: lowres.  This option can appear anywhere on the comand line and signals the local cluster regression to only run lowres (72 and 75 dpi) output, for example:
  clusterpush.pl pcl lowres

or

  clusterpush.pl lowres "gs pcl" henrys

Note that this will fail if we ever hire an engineer with the name lowres
or use that as an internal product name.  Also, there currently aren't
any 75 dpi baseline md5sums for the pcl product, so until the next pcl
svn commit it doens't make sense to run with this option.

Also added partial support for generating bitmaps with changes and running them through bmpcmp.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10663 a1074d23-0009-0410-80fe-cf8c14f379e6