More fixes to gs_2_colors.
It's important that the overprint device is updated when colorspaces change,
not just when the overprint setting changes. To this end we relax two of the
"has overprint changed" tests in the code. I'm not entirely sure how this
ever worked in trunk gs, but presumably there must be something else at play.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/gs_2_colors@10844 a1074d23-0009-0410-80fe-cf8c14f379e6