Type 3 font glyphs that use the d0 operator are free to set their own color as they run. They assume however that the current color is set in the graphics state when they are run.
Unfortunately, this doesn't map nicely into our "2 color" world, so we need
to ensure that Type3 d0 glyphs are always called with the current color set
both as stroke and fill color. The easiest way to do this is to set both
stroke and fill colors to the same thing before calling the show operator,
and that is what this review does.

This is a bit of a shame, because we'd really like to be able to have separate
colors when calling text stuff for pdfwrites sake. Will discuss this more
with Ken tomorrow to see if we can find a better solution.

This gets us down to just 5 differences, in 2 files.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/gs_2_colors@10903 a1074d23-0009-0410-80fe-cf8c14f379e6