Revert yesterdays commit 10903, in favour of a 'better' solution.
Rather than always setting stroke and fill colors to be the same when calling
text operators (which largely defaults the point of having them separate in
the first place) leave them be.

Alter the d0 operator to read the current color/colorspace and to set the
'alternative' color to be that. This fulfills the requirement that both
colors should be identical at the start of a Type3 glyph building procedure.

This leaves a varying number of differences in the local cluster tests,
varying from 0 to 3 files giving problems. The ones I have managed to
capture seem to involve glyphs being offset by a pixel from their expected
positions, and appear completely unrelated to any gs_2_color changes.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/gs_2_colors@10910 a1074d23-0009-0410-80fe-cf8c14f379e6