Fix for bug 691115.
In some circumstances, the shading code can get all the way to the point of
marking the screen/page without ever clipping to the relavent region. This
produces cases where we "draw outside the lines". Often this is masked by
the fact that we only ever draw as far out of the lines as would be covered
by an area of constant color (typically a small area).

The fix for this is to spot the unclipped case when we come to render
trapezoids to the screen, and to clip the trapezoid into (at most) 3
smaller trapezoids.

This causes 104 non-pdfwrite differences, and 24 pdfwrite ones. Examination
using localcuster bmpcmp shows them all to be acceptable.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10963 a1074d23-0009-0410-80fe-cf8c14f379e6