New Fontmap files for the URW fonts, both type 1 and TrueType formats.
The file Fontmap.URW-136.T1 contains definitions for the type 1 fonts (the URW PFB
fonts must be in the search path).

The file Fontmap.URW-136.TT contains definitions for the TrueType fonts (the URW TTF
fonts must be in the search path).

Currently neither of these work 100% because of problems with the URW font sets
(several incorrectly named fonts, and in one case an illegal type 1 font)

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10988 a1074d23-0009-0410-80fe-cf8c14f379e6