'Fix' FAPI with FreeType.
FreeType's hinting is worse than useless for type 1 fonts, its actually harmful. It
distorts the glyph shape because it hints an unscaled outline. While it would be best
to fix FreeType's hinting this is a large project. For now this patch disables FreeType
hinting when handling type 1 fonts.

This resolves many spacing issues as well as fixing the stroke/fill mismatch with
ia3z3519.pdf and the incorrectly scaled outline in womanface.pdf.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11009 a1074d23-0009-0410-80fe-cf8c14f379e6