Fix FreeType When using bitmaps from a TrueType/Type42 font FreeType uses a heuristic approach to synthesize vertical metrics for the glyphs, including a horizontal advance. As in the past, this works well as long as the client knows whether the advance direction is horizontal or vertical and uses the correct advance.
However we don't know that, so we don't want the vertical advance.

Its easiest jut to have FreeType ignore embedded bitmaps and use the outlines, in which
case we don't get the invented vertical advance. Not only that but the rendered bitmap
is a better match for the cases where we don't have a pre-built bitmap, so consistency
is improved with this patch as well.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11013 a1074d23-0009-0410-80fe-cf8c14f379e6