Fix for bug 691228, problems with wide strokes with round caps/joins with antialiasing enabled.
2 small problems in the same code. In the antialiased case, we stroke by
forming 2 paths (one up either side of the line) and then reverse one and
concatenate it. This should happen whenever we meet a closepath, but due to
a misplaced return, was being missed out in some round join cases.

Also, when using a round join in the case where 2 lines meet 'in-line' we
would inadvertently leave the point on the wrong side of the line.

No changes in localcluster testing as this only affects antialiased results
and that's not tested currently.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11031 a1074d23-0009-0410-80fe-cf8c14f379e6