Remove global variables in gdevcdj.c as part of the effort for bug 691207.
The ESC/P based printing code in gdevcdj.c uses various global variables; we
gather them into a struct here, and pass the struct between the 2 functions
concerned.

This device isn't used in the default build, so no point in testing with the
localcluster. Enabling it locally and doing before/after tests shows that
identical results are given for rendering tiger.eps.

A side effect of this commit is to remove the nasty shadowing of the img_rows
variable (doubly nasty because of it's use within a macro). This meant that I
had the code wrong for a while, so I know the test is actually testing the
right thing.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11036 a1074d23-0009-0410-80fe-cf8c14f379e6