Remove the function scope static variables from gdevcdj.c as part of the efforts for bug 691207 (multithreaded safety). These only affect the ESC/P driver.
There are 4 variables in play here. real_rows doesn't actually need to be
static at all as it is always set before being used. The other 3 (ln_idx,
vskip1 and vskip2) all go into the 'ep_globals' structure.

There is a subtle change in behaviour here; previously these variables were
initialised to zero on execution start (and again on every flush). Now they
are initialised to zero at the start of every page (and again on every flush).
This means that leftover vskips are no longer carried forwards from page to
page. This seems like a good thing to me.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11038 a1074d23-0009-0410-80fe-cf8c14f379e6