Remove the commented out invocations of autoheader and aclocal/automake in the autotools bootstrap script.
The various autotools modules must be invoked in a specific order, and
usually with non-default options. We only use autoconf in our build,
not any of the other tools which layer on top of it, but I had left
their invocations in the script, but commented out, so make it easier
to add them at a later date.

It's been some years since autogen.sh was added, and we still have no
plans to adopt automake. Also, the script didn't include support for
libtoolize (which is glibtoolize of MacOS X)  nor for detecting the
appropriate automake version, so it seems simpler to just remove the
commented out lines to avoid confusion.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11042 a1074d23-0009-0410-80fe-cf8c14f379e6