Modify FAPI support files Moved FAPIConfig, FAPIcidfmap and FAPIfontmap from the 'lib' directory to the 'Resource/Init' directory, so that they get built in when COMPILE_INITS=1.
Also modified gs_fapi.ps to contain default values for the contents of FAPIConfig which
map all fonts to the FAPI engine (FreeType for now). Modified FAPIConfig so that the
usual contents are commented out. This should allow the possibility of overriding the
default values by putting new values in FAPIConfig.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11059 a1074d23-0009-0410-80fe-cf8c14f379e6