Fix FreeType FAPI No bug for this (probably should be). Noticed while testing that using a TrueType to replace a missing CIDFont (ie entry in cidfmap) was rendering the wrong glyphs, and applying a vertical advance.
The incorrect glyph seems to be because we were incorrectly stating that the character
index was not a glyph ID, which it is in this case.

The second problem was more difficult. FreeType manufactures vertical advance metrics
for all TrueType glyphs, we override those in the incremental interface callback, but
when the TT is a CIDFont on disk, we don't use the incremental interface and so have
no opportunity to override the metric.

Addressed this by checking in load_glyph to see if the current font is not type 1
(ie its TT), *is* a CIDFont, and is not a vertical font (WMode is 0). In this case
we simply override the vertical advance and maek it 0.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11070 a1074d23-0009-0410-80fe-cf8c14f379e6