Made oog_encode in tif_luv.c reentrant by precalculating oog_tables.
oog_tables was calculated on-demand when writing TIFF files in LogLuv Encoding.
Since it didn't depend on anything dynamic, I replaced this code with a static
const array.  The array only consisted of 100 ints - the net library size
actually went down a few bytes after this change.

This is part of bug #691207 (global variables must be removed).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11088 a1074d23-0009-0410-80fe-cf8c14f379e6