Simplify ZIP parsing.
The special mode of the XPS interpreter that parses the
ZIP file incrementally as a stream has proven to
be useless. None of the files we see in the wild make
use of the features that are required for a streaming
XPS parser to behave well. As a result, this mode was
always buffering up data and using lots of memory.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11130 a1074d23-0009-0410-80fe-cf8c14f379e6