Move op_array_table_global and op_array_table_local away from being globals and into i_ctx_p. This requires various knock on changes throughout the code to enable them to be recovered from the context (which may itself need to be recovered from a memory pointer).
Also, the garbage collection is altered slightly; rather than the
op_arrays being special cased (as special objects that have to have
their component parts registered as roots for the garbage collection) we
move them inline into i_ctx_t and amend the gc routines for that
structure. Thanks to Alex for his help with this.

No differences expected, or shown with local cluster testing.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11146 a1074d23-0009-0410-80fe-cf8c14f379e6