Fix (pdfwrite) incorrect ToUnicode CMap. Bug #691274 "Missing or incorrect ToUnicode when using Identity ordering"
pdfwrite emits ToUnicode CMaps in PDF files which always use 2 bytes for the key
length. That is 0x21 is represented as <0021>, since these are hex strings there should
in fact be no difference between <21> and <0021>. It is clear from the ToUnicode
Mapping file tutorial that this is correct for mapping files.

However, a PDF file is not a CMap file. It appears that Acrobat does not like
beginbfrange data which is padded with zeros (it seems happy with codespacerange and
destination codes which are 0 padded though).

This patch outputs ToUnicode CMaps which are not padded with 0, ie 0x21 is emitted as
<0021>, for the srcCodeLo and SrcCodeHi parameters in a beginbfrange operation. The
destination codes remain zero-padded.

Technically this would be incorrect for a real CMap file, but we don't write those
out so this should not be an issue. The change is limited to ToUnicode CMaps and does
not affect any other kind of CMap.

I believe the change will affect the data emitted for the beginbfrange,
beginbfchar, begincidchar, beginnotdefchar, beginnotdefrange and begincidrange
operators. It does *not* affect the begincodespacerange operator; although Acrobat
Distiller writes this without zero padding tests show that Acrobat is happy reading
zero-padded code space ranges.

ToUnicode information is not used in rendering PDF files, so regression testing should
show no differences. Local testing of a limited number of files doesn't reveal any
problems, but this change has not been well tested.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11170 a1074d23-0009-0410-80fe-cf8c14f379e6