Fix (FAPI) The PDF interpreter, when a font is missing, and it has the 'small caps' flag set, creates a substitute by altering the CharStrings of the substitute font, replacing the lower case glyph CharStrings with procedures which 'show' the upper case glyph at a reduced size.
This ends up in the FAPI code as a t_mixedarray type which was being missed by the check
for a t_array, added previously to handle arbitrary procedures in a type 1 font. After
extending the check to include t_mixedarray it was clear that the SEAC processing was
also messing up the dictionary lookup, so it was necessary to restore a value which
had been overwritten in case a SEAC turned up.

zero-length-font.pdf should now work as it does without FAPI, no other differences
expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11220 a1074d23-0009-0410-80fe-cf8c14f379e6