Fix FAPI interface There is code in the FAPI interface for replacing a font's /FontBBox entry, this sometimes has to create a new array to insert into the dictionary, if the old array is not suitable for storing real values.
The code to do this continued using ref_assign_old, as it had done for replacing the
values in an existing array. This appears to be incorrect, when created the refs in a
new array are uninitialised so attempting to use ref_assign_old tries to use
uninitialised data.

Modified to use ref_assign_new instead, as indicated by the comments in store.h

This *may* reduce some of the non-deterministic behaviour, possibly with files such
as ShadowText.eps

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11232 a1074d23-0009-0410-80fe-cf8c14f379e6