The final part of trying to "ln -s base/unix-gcc.mak Makefile ; make so" work.
There were some fundamental problems with the older version: The so target really
should be built in three distinct stages with three different FLAGS - the
utils (genarch, gechogs) executable part, the shared library part, and the two final
executable part. The library should be built with the shared-object flags, but neither
of the utils nor the final executables should be built with "-shared", since they are
executables and not libraries. Hope this is clear.

There is also a little problem from SOC_LOADER not being defined if this file
is not used from ./configure's Makefile.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11257 a1074d23-0009-0410-80fe-cf8c14f379e6