Some work towards (but not a complete cure for) bug 691030.
Previously XXX_slant_into_trapezoids would check for 2 'simple' cases,
and bale into the fully general fill_slant_adjust function otherwise.
The simple cases were where the top span was a subset of the bottom one, or
vice versa.

Unfortunately the way the code to test for these cases was originally written,
one of these special cases could be used for shaped with vertical sides, and
the other couldn't.

By rewriting the test slightly, both cases can be used with trapezoids that
have one vertical side.

For good measure we also extract the 'rectangle' case out separately.

Cluster testing shows no differences (barring a couple of indeterminisms).

This reduces the output from 5 shapes (2 rectangles, 3 traps) to 3 (2
rectangles, 1 trap). One of the rectangles is still a repeat, but it's
a step in the right direction.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11282 a1074d23-0009-0410-80fe-cf8c14f379e6