Fix copy CIDFontType 2 properly Bug #690930 "Unicode Output: DEVICE=pdfwrite"
When making a copy of a CIDFont with TrueType outlines (CIDFontType 2) for the
benefit of pdfwrite, do not simply copy the subst_CID_on_WMode structure member as
this will be restored away along with the font at end of page. Instead make a new
structure and store that in the copied font.

No differences expected, local cluster test shows a few but I think (hope) these are
related to the ICC branch merge.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11321 a1074d23-0009-0410-80fe-cf8c14f379e6