Fix pdfwrite, faulty matrix transformations for Shading dictionaries
Bug #691352 "cairo pdf mis-distilled"

Patterns in PDF are unpleasantly complicated by the need to transform the pattern to
the 'default co-ordinate space'. Normally this means that we undo the resolution scaling
which is normally applied to the CTM.

For page streams this works well, but for forms the 'default co-ordinate space' is
the space of the parent. For one level of form there is therefore no difference between
the page and the form. When forms are nested however, the lower form's space becomes
that of the parent. This means that patterns inside forms, which are nested inside
another form, need to be transformed to the parent form co-ordinate space, not the
page space.

Since we don't currently emit forms from pdfwrite for anything except transparency
groups what this means in practice is that we don't undo the CTM transformation if
we are rendering a pattern inside a form, nested inside at least one other form.

Expected Differences
These files show progressions with this change:
Buig690831.pdf
Bug690208.pdf
Bug690206.pdf
Bug690115.pdf
Catx5233.pdf

These files show progressions, but also show regressions or are still not properly
converted to PDF:
Bug688807.pdf
Bug689918.pdf

This does not conclude the work for bug #691352, and the two files which still exhibit
issues also require further work.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11347 a1074d23-0009-0410-80fe-cf8c14f379e6