Fix (pdfwrite) : Don't emit invalid PDF/A files
Bug #691383 "pdfwrite emits invalid PDF/A when glyphs are missing in the input font."

If the input file references glyphs which are not in the font, then we could emit a
PDF/A file which also referenced glyph which were not present in the fonts embedded
in the PDF file.

This patch uses the PDFACompatibilityPolicy to control the action in this case, and
avoid emitting invalid PDF/A files.

Also extended the functionality of the PDFACompatibilityPolicy to allow the user to
abort PDF file production when a valid PDF/A file cannot be produced.

No expected differences as PDF/A production is not tested by regression code.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11369 a1074d23-0009-0410-80fe-cf8c14f379e6