Fix (pdfwrite) : limit composite object sizes when producing PDF/A
Bug #691370 "Dictionary size not limited for PDF/A-1"

The PDF/A specification insists that the implementation limitations of Acrobat as
defined in PDF 1.4 should be followed when producing PDF/A files. Accordingly this
limits the size of dictionaries to 4095 and arrays to 8191 when producing PDF/A output.

Since there is no way to ignore the problem, setting the PDFACompatibiltyPolicy to 1
'don't emit the object, carry on producing PDF/A' can't be made to work. Since the
user has expressed a desire for an output file, we fall back to Policy 0 'Produce a
regular PDF file' instead.

Also extend PDFACompatibilityPolicy to include policy 2 'abort'.

No test file for this case, no differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11372 a1074d23-0009-0410-80fe-cf8c14f379e6