Enhancement (pdfwrite) : Fall back to Flate instead of JPEG when an image turns out short.
Bug #691435 "any PDFSETTINGS besides /default gives dictfull error"

There are two problems; firstly when an error is encountered closing a filter the
error code (-1 or -2 usually) is returned and interpreted as a regular GS error,
resulting in ridiculous error messages. Modified to return ioerror instead.

Secondly, when an image terminates with insufficient data and pdfwrite is creating a
DCT encoded image, we can't complete the image. In this case try to fall back to the
simple Flate compressed stream instead.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11476 a1074d23-0009-0410-80fe-cf8c14f379e6