Fix (pdfwrite) : Named destinations with no /Page not working properly
Bug #691344 "REGRESSION: ps2pdf of 8.70 generates correct hyperlinks but not ps2pdf of
8.71"

The revisions 9779 and 9788 fixed a problem where DOCVIEW pdfmarks without a /Page or
/View entry were producing invalid PDF files, and a problem with DEST pdfmarks arising
from it.

Because DOCVIEW and DEST pdfmarks share some common code in pdfwrite to create a
Destination annotation in the outptu PDF, this still left some confusion in the code.
The problem was that DOCVIEW pdfmarks need not contain Page or View data, but named
destinations *must*, but pdfmarks need not specify, in which case the current page is
used.

Added a parameter to pdf_make_dest which allows us to differentiate the two cases and
use the current page number for a named destination if no page is supplied.

The regression suite can't test this non-printing feature, so no differences are
expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11491 a1074d23-0009-0410-80fe-cf8c14f379e6