A much better solution for Bug 691328.
The bitmap data coming from a font scaler/renderer may not have its raster
alignment as Ghostscript expects, so in the event that is the case, create
a bitmap which is GS compatible in the FAPI layer, before calling any
further operations on it.

This reverts the previous "hack" from r11362.

No cluster differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11542 a1074d23-0009-0410-80fe-cf8c14f379e6