Two changes for two primarily pdfwrite problems with FAPI/FT.
The first is to only replace whichever of BuildGlyph or BuildChar
procedures the original font actually contained, replacing both
regardless confused some of "decoding" code elsewhere in GS
(op_show_continue_dispatch()) and we ended up receiving a valid
glyph to render instead of notdef. In this case we ended up incorrectly
stroking a charpath (this was seen with Quark's Type 3 "special effects"
font). With rendering output this resulted in a boldening effect on the glyph,
in pdfwrite, the result was strange extra lines through the glyph.

The second change is to due pdfwrite's use of a unit matrix for some of
its work. This resulted in small glyph metrics rounding to zero within
Freetype, and pdfwrite thinking a glyph had a width of zero. So make
the scale value at which we consider a glyph to be "small" an order of magnitude
larger, which gets compensated for in the calculation of Freetype's final
matrix.

Unfortunately this does cause ~2600 differences in the cluster tests: some
are progressions, most are pixel differences due to rounding differences
because the scaling numbers have changed in Freetype.

Fixes Bug 691383

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11549 a1074d23-0009-0410-80fe-cf8c14f379e6