Fix (pdfwrite) : try not to write out invalid xref
Exposed by bug #691503. There are circumstances under which we create a reference to a
/Page object which does not yet exist (creating an outline tree for example). In this
case we create a reference to the page, and set its initial offset to the current
position in the PDF file. When we later write the real object we update the position
to reflect the point in the file where it is actually stored.

However, if we don't actually ever write the object (in the case above, because
LastPage is set before the destination page), then the object remains pointing to the
wrong location in the file. This is then written into the xref table, and this confuses
the current version of Ghostscript.

The patch creates new pages with an offset of 0 (which is always invalid because the
PDF header means the minimum offset for any object is 15). If the page is later
created then the offset gets updated. If it is not written, however, we now check in
the xref writing code to see if an object has an offset of 0. If it does, then we
break the xref table up and do not write an xref entry for the missing object.

Ghostscript is happy with this and processes the file.

NB the proper fix is to not emit the Outline entry for the page which will not be
present in the output, and a patch has been suggested to achieve this. This code has
some uses, however, as it may prevent this kind of problem in the future. It also
permits objects to be created and never referenced (presumably because they are never
used) without leaving invalid entries in the xref, which might be useful.

No expected differences.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11550 a1074d23-0009-0410-80fe-cf8c14f379e6