Assorted fixes for the modularized X11 output device
All object files destined for shared libraries needed to be compiled
with -fPIC but this is not the case for gsparamx.c, gdevemap.c, and
gdevxcmp.c which are required by other parts to be built as standard.

This change makes x11 not depend on gsparamx.c and gdevemap.c (and not
geeting them included in the shared library X11.so). gsparamx.c is
part of psdf and is usually included and available, and gdevemap.c is
small enough to be inserted completely where it is used.

gdevxcmp.c is addressed by building it with $(GLCCSHARED) and not with
$(GLCC).

In addition, $(CC_SHARED) is only used for object files, rather than
for linking, so DYNAMIC_CFLAGS is more appropiate here than
DYNAMIC_LDLAGS.

Bug 691510

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11555 a1074d23-0009-0410-80fe-cf8c14f379e6