Fix for Bug 691504 (gs taking a long time to render a file).
Revision 8694 changes code from using a simple cast to an int to using
a call to (int)floor. This fixes problems, but costs a factor of 3 in
processing time for heavily tiled files.

Here we change to using a macro version of floor instead. This removes
the function call and NaN handling overheads and returns us to the
same kind of performance as we had before.

We apply the same change throughout the file (i.e. all the calls to (int)floor,
not just the two changed in r8694.

There is still scope for optimising this code further, but this closes the
regression.

No differences seen in local cluster testing.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11599 a1074d23-0009-0410-80fe-cf8c14f379e6