Fix (pdfwrite) : faulty matrix transformations for Pattern dictionaries
Bug #690398 "Geometric issue with pattern"
The problem is the definition of the default co-ordinate space. When a Pattern is at the
page level or drawn inside a form, we need to remove our 0.1 scaling in order to return
to the 'default co-ordinate space'. However, when painted on a form nested inside a
form, the default co-ordinate space is the parent form, so we don't need to undo
the scaling.

This is the same problem addressed for shading dictionaries in revision 11347.

Expected Differences
Bug6901014_CityMap-evince.pdf
Bug6901014_CityMap-evince.simplified.pdf
Bug689422.pdf

All show progressions with this change.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11662 a1074d23-0009-0410-80fe-cf8c14f379e6