More Windows batch file madness.
Guard all path expansion with quotes, since the Windows command shell processor is too
stupid to do so, even when expanding variables/arguments itself.

Add a trailing separator to the %TEMP% variable before specifying the filename, as
apparently some setups don't get a trailing separator by default. A paired separator
(ie \\) doesn't seem to cause a problem so this ought to be safe even if the setup
does have a trailing separator.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11684 a1074d23-0009-0410-80fe-cf8c14f379e6