Slight refinement in the how the FAPI character generation detects we are producing anti-aliased glyphs.
Also, it seems that when assigning variable value to a second, larger data
type, whilst gcc promotes the data type before the assignment, MSVC promotes
the variable later, so shifting a 32 bit integer up during an assignment to
a 64 bit integer variable works find on gcc, but MSVC loses the most
significant bits. So we need to force the promotion to happen with a cast.

Bug 691588

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11695 a1074d23-0009-0410-80fe-cf8c14f379e6