Fix (pdfwrite) : incorrect pattern matrix calculation
Bug #690504 "ps2write does not work with asymmetric resolutions, like 600x1200dpi "

The problem was caused by frankly totally incorrect calculations when generating the
/Matrix for a pattern. Because we always see the CTM including the resolution scaling
but need to output a matrix with this removed, we scale the CTM by the inverse of the
resolution scaling.

But the calculation simply divided the CTM through by the scale factors, which is
totally incorrect and fails badly if the matrix involves rotation. The code now
correctly creates and multiplies a scale matrix. Note, the order of multiplication is
important here.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11699 a1074d23-0009-0410-80fe-cf8c14f379e6