As part of the investigation into bug 691612, we noticed that the clist band playback code doesn't call the device stroke_path method, but instead calls the default implementation directly.
In order the fix the bug, we'd like it to call through the device version
(pdf14_stroke_path). Consensus is that it's an oversight that it's not calling
the device version already, and tests indicate no differences (*) so this
patch changes it to call via the vector.

(* Actually regression testing shows 2 pdfwrite differences, but they are
in unbanded tests, so can't be executing this code, so are probably just
indeterminisms).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11737 a1074d23-0009-0410-80fe-cf8c14f379e6