Final part of the fix for bug 691612.
Currently the code blindly sets the "lop_pdf14" whenever the pdf14 device is
in use. This forces the 'lop_is_idempotent' test to fail, which in turn forces
slower cases to be taken through the code.

This patch alters that logic, so that the pdf14 device can leave that bit
unset if the parameters are set up in such a way that the lop really can
be idempotent (i.e. normal blending, solid alphas etc).

This vastly improves the speed of rendering non-alpha content, so we are no
longer penalised for having just a small amount of transparent content on a
page.

Testing shows 398 files that change, but bmpcmp reveals these all to be very
small changes to do with whether individual pixels are filled or not. This
is consistent with the typical mismatches between special case and normal
renderings, so can safely be ignored for now.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11738 a1074d23-0009-0410-80fe-cf8c14f379e6