Fix (TrueType fonts) : memory leak
Bug #691088 This does not affect the PostScript/PDF interpreters which now use FreeType
but is relevant still for XPS and PCL.

A buffer was allocated to contain the GSUB table from the TT font, but was never freed
leading to memory leaks. I've adopted the same approach as that taken for the 'glyph
length' array and added a notification procedure which is called when the font is
released, and that procedure frees the GSUB buffer.

At the same time, removed the functions:
add_quadratic_curve
append_simple
check_component

These were only called by each other, or by append_component which had already been
removed as unused.

No expected differences.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11741 a1074d23-0009-0410-80fe-cf8c14f379e6