Fix (pdfwrite/ps2write) : Don't dereference mask colour spaces
Bug #691106 "-sDEVICE=pdfwrite -dConvertCMYKImagesToRGB=true"

When checking to see if an image needs to be converted to RGB, the image space is
checked to see if it is CMYK. However an image mask does not have a colour space (it
is NULL), firstly so that we can tell its an image mask, and secondly because masks
genuinely do not have a colour space, they take on the current colour.

Attempting to dereference the NULL colour space causes a crash. Since a mask never
needs converting to RGB, we short circuit the test by detecting a NULL colour space.

Expected Differences
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11744 a1074d23-0009-0410-80fe-cf8c14f379e6