Fix (Fonts): Prevent xfont being used in composefont
On platforms supporting X-Windows xfonts can be used with Ghostscript, however we don't
want to find an xfont (eg Courier) when using composefont to create a CID-Keyed instance
as this will not work.

Test to see if the top level font is a type 0, and if it is, don't search for a matching
xfont, but instead go through the usual PostScript machinery.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11745 a1074d23-0009-0410-80fe-cf8c14f379e6