Fix for part 1 of bug 691484 - uninitialised memory usage in clist code.
The FAPI code copies bitmaps and pads them out so that each raster line
is appropriately aligned. Previously the padding bytes were left
uninitialised. Unfortunately the clist code reads the padding when
compressing the bitmaps, which a) upsets valgrind and b) means that the
exact size of the clist will vary from run to run.

Here we tweak the copying code to ensure that the padding bytes are always
set to zero.

No differences shown in local cluster. This solves part 1 of the bug.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11759 a1074d23-0009-0410-80fe-cf8c14f379e6