Fix a potential NULL pointer dereference flagged by static analysis.
Other potential occurrences flagged by the analyser in this module are deemed not to be
possible, as these are picked up at a higher level.

For instance, dereferencing a pointer to a path structure in a path handling method. We
can assume the path pointer is not NULL as we would not be called if the path was empty.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11767 a1074d23-0009-0410-80fe-cf8c14f379e6