Put back some statements. It turns out that we can either have gcc be warning free or the static analysis warning free. The problem is gs_note_error which uses gs_log_error, on a release build gs_log_error is defined as the error value.
If we don't do something with that value then gcc complains that the 'statement has no
effect'. So the code was set like this:

ecode = gs_note_error(error code);

But if we want to actually ignore the error and just note the problem, then we don't use
ecode and the static analysis complains that the variable is unused....

I'd rather have no warnings from gcc so I've restored that.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11771 a1074d23-0009-0410-80fe-cf8c14f379e6