Fix (pdfwrite) : Check encoding compatibility when finding base fonts
Bug #691656 "crash in Chinese font: /typecheck in --.FAPIBuildGlyph--"

A typecheck error is not a crash of course, but when using the built in font interpreter
this did not produce an error, silently discarded the text instead. This was still
incorrect.

The problem is caused when dealing with type 0 fonts and producing multiple type 1 fonts
to handle the large number of glyphs. The code for finding an existing font (in order
to minimise the number of new fonts) or create a new font, did not consider the
encoding of an existing font when trying to match it (this did work for regular fonts).

This could lead to a font with an incompatible encoding being used, which caused an
error later in text processing where a routine is supposed to be guaranteed no font
encoding problems. That led to an attempted fallback to the 'bitmapped type 3 font'
solution, but the text processing was passing ridiculous values to the font interpreter.

In the old font code this caused a silent discard of the text, with the FreeType code
it produces an error.

Fixed by checking the base font we find to see if its encoding is compatible with the
current text encoding, and manufacturing a new font if it is not.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11774 a1074d23-0009-0410-80fe-cf8c14f379e6